Move TestInstantiateStructuredConfigs to test_to_container.py
